Relocate department and member item
