+++ frontend-next-migration/src/entities/Member/ui/DepartmentItem.module.scss
@@ -1,0 +1,3 @@
+.departmentCard {
+  margin-bottom: 2rem;
+}
