+++ frontend-next-migration/src/entities/Member/ui/MemberItem.module.scss
@@ -1,0 +1,67 @@
+.memberRow {
+  display: flex;
+  width: 100%;
+  justify-content: space-between;
+
+  .memberName {
+    font-weight: bold;
+    text-align: end;
+    flex: 2;
+    min-width: 0;
+    padding-right: 2rem;
+    display: flex;
+    justify-content: flex-end;
+  }
+
+  .taskText {
+    text-align: left;
+    flex: 2;
+    min-width: 0;
+  }
+
+  .iconContainer {
+    display: flex;
+    gap: 0.5rem;
+    align-items: center;
+    justify-content: end;
+    flex: 1;
+    min-width: 0;
+  }
+
+  .centerContainer {
+    display: flex;
+    flex: 2;
+  }
+
+  .clickableLogo {
+    color: var(--secondary-color);
+    cursor: pointer;
+    @include link-scale-effect('icon');
+  }
+
+  .memberLogo {
+    width: clamp(0.5rem, 3vw, 2rem);
+    height: clamp(0.5rem, 3vw, 2rem);
+    border-radius: 50%;
+    overflow: hidden;
+    @include link-scale-effect('icon');
+  }
+
+  .Logo {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    border-radius: 50%;
+    cursor: pointer;
+    @include link-scale-effect('icon');
+    background-color: var(--secondary-color);
+  }
+
+  .placeholderLogo {
+    width: 100%;
+    height: 100%;
+    opacity: 0;
+    border-radius: 50%;
+    @include link-scale-effect('icon');
+  }
+}
